+++ src/Enum/DecisionStrategyUnanimous.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Fschmtt\Keycloak\Representation;
+namespace Fschmtt\Keycloak\Enum;
 
 class DecisionStrategyUnanimous extends DecisionStrategy
 {
