+++ src/Enum/Enum.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fschmtt\Keycloak\Enum;
+
+abstract class Enum implements \Stringable
+{
+    abstract public static function from(string $value): static;
+}
