--- src/Representation/DecisionStrategy.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Fschmtt\Keycloak\Representation;
-
-abstract class DecisionStrategy extends Representation implements \Stringable
-{
-}
