--- src/Representation/PolicyEnforcementMode.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Fschmtt\Keycloak\Representation;
-
-use Stringable;
-
-abstract class PolicyEnforcementMode implements Stringable
-{
-}
