+++ tests/Unit/Enum/DecisionStrategyAffirmativeTest.php
@@ -2,14 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Fschmtt\Keycloak\Representation;
+namespace Fschmtt\Keycloak\Enum;
 
 use PHPUnit\Framework\TestCase;
 
-/**
- * @covers \Fschmtt\Keycloak\Representation\DecisionStrategyAffirmative
- * @uses \Fschmtt\Keycloak\Representation\Representation
- */
 class DecisionStrategyAffirmativeTest extends TestCase
 {
     public function testConvertsCorrectlyToString()
