+++ tests/Unit/Enum/DecisionStrategyUnanimousTest.php
@@ -2,14 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Fschmtt\Keycloak\Representation;
+namespace Fschmtt\Keycloak\Enum;
 
 use PHPUnit\Framework\TestCase;
 
-/**
- * @covers \Fschmtt\Keycloak\Representation\DecisionStrategyUnanimous
- * @uses \Fschmtt\Keycloak\Representation\Representation
- */
 class DecisionStrategyUnanimousTest extends TestCase
 {
     public function testConvertsCorrectlyToString()
